Fix versioning problems and upgrade build tools
Imports of `ddnswolf` in setup.py were causing builds to fail, as well as imports of `git` in ddnswolf/version.py. These imports have been reorganized to reduce problems and keep the builds running.

An additional VersionInfo class was created, SimpleVesionInfo, to supply only the release number. This class is used when `git` cannot be imported, to allow the build to continue in any situation.

calculate_full_version() was also renamed to get_full_version(), simply as a consistency change since that name is already used in each individual class.

Lastly, a small change to the Arch PKGBUILD. Dashes in our version string are now underscores instead of dots, as dots were confusing to look at when our release number also has dots.