Improve Arch PKGBUILD
Local source copying is now a bit less janky by specifying a different
source in the source array depending on LOCAL_SOURCE's value.

Previously, the public git repo would be downloaded regardless of if a
local source was being used. Now that will not happen. This also removes
the need to cd into a separate source directory during build.

Lastly, various other fixes were made at the direction of the `namcap`
tool provided by ArchLinux. Maintainer comment was changed to the
standard form, and `git` was added as a dependency which is required
for any package that downloads sources from git. The _src_pyroot()
function was also removed since it is discouraged to define additional
private functions.