Refactor DNS update protocols, rename to 'updaters'
They have been renamed from Protocols to Updaters to better reflect
their purpose. The sources can also make use of a "protocol",
referring to a specific service or API. It is less confusing to
call these modules "updaters" since they specifically perform a DNS
update.

And as a final note, I renamed the abstract class "DynamicDNSUpdater"
to be just "DNSUpdater" because it looked better.