Fix bug when running dev version outside git repo
If the working directory was not within the DDNSWolf git repo,
the DynamicVersionInfo constructor would fail with a
"repository not found" exception. This was not caught, and the
program would crash. The fixed behavior is to also catch that
exception and fall back to SimpleVersionInfo, just like if the
`git` python module was missing.