Fix bug where all filters were PublicAddressFilter
The conversion code for address subscription filters
had a bug that caused every filter name in the spec
to be evaluated as an instance of PublicAddressFilter,
rather than their own class.

This was caused by an improper function closure.
create_this_filter() (now renamed create_filter_local) used
the local variable `filter_cls` to invoke the constructor for
the filter class currently selected. This variable was not
copied for each instance of the function - each instance
referenced the same value.

This variable is modified for each iteration,
and the final value of this variable once the loop exits is
`PublicAddressFilter`. That particular class is last because
it is last in the import order. Thus, when the loop is done,
and the filter specifications are evaluated, every call to
create_this_filter() would use the stale value of filter_cls.

This bug is fixed by adding a new nested function filter_local()
which serves as the function local used in eval(), and the
outer function serves only to `return` the inner one with its
unique filter_cls. The return is required to create a true
python closure.